Run function nestor on a list of initial cliques using parallel computation (mclapply)

Run function nestor on a list of initial cliques using parallel computation (mclapply)

List.nestor(
  cliqueList,
  Y,
  sigma_O,
  MO,
  SO,
  r,
  alpha = 0.1,
  cores = 1,
  maxIter = 20,
  eps = 0.001,
  trackJ = FALSE
)

Arguments

cliqueList
a list containing all initial cliques to be tested

Y
count dataset

sigma_O
result of PLN estimation: variance covariance matrix of observed data

MO
result of PLN estimation: means matrix of observed data

SO
result of PLN estimation: marginal variances matrix of observed data

r
number of hidden variables

alpha
tempering parameter

cores
number of cores

maxIter
maximal number of iterations of the algorithm

eps
convergence precision parameter

trackJ
boolean for the estimation of the lower bound at each parameter update

Value

a list containing the fit of nestor for every cliques contained in cliqueList

Examples

 data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- find a list of initial cliques
findcliqueList=boot_FitSparsePCA(data$Y, B=5, r=1)
cliqueList=findcliqueList$cliqueList
length(cliqueList)

#> [1] 1

#-- run List.nestor
fitList=List.nestor(cliqueList,data$Y, sigma_O, MO,SO, r=1)
length(fitList)

#> [1] 1

str(fitList[[1]])

#> List of 3
#>  $ message: chr "could not find function \"nestor\""
#>  $ call   : language nestor(Y = Y, MO = MO, SO = SO, initList = init, eps = eps, alpha = alpha,      maxIter = maxIter, verbatim = 0, | __truncated__
#>  $ clique :List of 1
#>   ..$ : int [1:2] 3 7
#>  - attr(*, "class")= chr [1:3] "simpleError" "error" "condition"