Core function of nestor.

Core function of nestor.

nestor(Y, MO, SO, initList, maxIter = 20, eps = 0.01, alpha = 0.1,
  verbatim = 1, print.hist = FALSE, trackJ = FALSE)

Arguments

Y
count dataset

MO
estimated means from norm_PLN

SO
estimated marginal vairances from norm_PLN

initList
result from initVEM

maxIter
maximun number of iterations

eps
convergence precision parameter

alpha
tempering parameter, default to 0.1

verbatim
controls verbosity, between 0,1 and 2.

print.hist
prints edges weights histograms at each step if TRUE

trackJ
computes the lower bound at each parameter update if TRUE. Otherwise, the lower bound is only computed at each new VE step.

Value

M estimated means

S estimated marginal variances

Pg edges probabilities

Wgedges variational weights

W edges weights

Omega market matrix filled with precision terms common to all spanning trees

lowbound table containing the lowerbound trajectory

features table containing the parametes trajectory

finalIter number of iterations until convergence was reached

time running time of the VEM

max.prec boolean for reach of maximal precision reached during the VEM fit

Examples

data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)

#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!

MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_obs=PLNfit$sigma_obs
#-- initialize with true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(data$Y,cliqueList=initClique,sigma_obs, MO,r=1 )
#-- run core function nestor
fit=nestor(data$Y, MO,SO, initList=initList, maxIter=5,verbatim=1)

#> 
#> nestor ran in 0.331secs and 5 iterations.

str(fit)

#> List of 11
#>  $ M        : num [1:100, 1:11] -0.119 0.896 0.542 -0.36 0.185 ...
#>  $ S        : num [1:100, 1:11] 0.0726 0.0384 0.0492 0.0816 0.0615 ...
#>  $ Pg       : num [1:11, 1:11] 0.00 1.00 0.00 0.00 4.06e-13 ...
#>  $ Wg       : num [1:11, 1:11] 0.00 1.29e+14 0.00 1.06e-06 1.84e+01 ...
#>  $ W        : num [1:11, 1:11] 0.00 1.01e+09 0.00 0.00 2.04e-04 ...
#>  $ Omega    : num [1:11, 1:11] 3.356 1.316 -0.479 -0.534 -0.889 ...
#>  $ lowbound :'data.frame':	6 obs. of  5 variables:
#>   ..$ J        : num [1:6] -579 -572 -571 -571 -570 ...
#>   ..$ T1       : num [1:6] -864 -885 -888 -888 -887 ...
#>   ..$ T2       : num [1:6] -4.16e-02 -2.30e-04 -1.90e-06 -3.32e-08 -1.16e-09 ...
#>   ..$ T3       : num [1:6] 284 313 316 317 317 ...
#>   ..$ parameter: chr [1:6] "complete" "complete" "complete" "complete" ...
#>  $ features :'data.frame':	5 obs. of  4 variables:
#>   ..$ diffPg   : num [1:5] 5.00e-01 1.39e-02 1.65e-04 1.08e-06 3.04e-08
#>   ..$ diffW    : num [1:5] 2.81e+04 3.86e+06 2.23e+07 8.82e+08 1.59e+10
#>   ..$ diffOmega: num [1:5] 0.6796 0.0524 0.0361 0.0346 0.0272
#>   ..$ diffWg   : num [1:5] 2.99e+12 2.72e+08 2.01e+11 1.77e+12 8.99e+13
#>  $ finalIter: num 5
#>  $ time     : 'difftime' num 0.330919027328491
#>   ..- attr(*, "units")= chr "secs"
#>  $ max.prec : logi FALSE