Core function of nestorFit

Core function of nestorFit

nestorFit(
  Y,
  MO,
  SO,
  initList,
  maxIter = 20,
  eps = 0.01,
  alpha = 0.1,
  verbatim = 1,
  print.hist = FALSE,
  trackJ = FALSE
)

Arguments

Y
count dataset

MO
estimated means from norm_PLN

SO
estimated marginal vairances from norm_PLN

initList
result from initVEM

maxIter
maximun number of iterations

eps
convergence precision parameter

alpha
tempering parameter, default to 0.1

verbatim
controls verbosity, between 0,1 and 2.

print.hist
prints edges weights histograms at each step if TRUE

trackJ
computes the lower bound at each parameter update if TRUE. Otherwise, the lower bound is only computed at each new VE step.

Value

M estimated means

S estimated marginal variances

Pg edges probabilities

Wgedges variational weights

W edges weights

Omega market matrix filled with precision terms common to all spanning trees

lowbound table containing the lowerbound trajectory

features table containing the parametes trajectory

finalIter number of iterations until convergence was reached

time running time of the VEM

max.prec boolean for reach of maximal precision reached during the VEM fit

Examples

data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- initialize with true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(data$Y,cliqueList=initClique,sigma_O, MO,r=1 )
#-- run core function nestorFit
fit=nestorFit(data$Y, MO,SO, initList=initList, maxIter=5,verbatim=1)

#> 
#> nestor ran in 0.303secs and 4 iterations.

str(fit)

#> List of 11
#>  $ M        : num [1:100, 1:11] 2.3162 -0.5346 1.4843 0.0824 0.7946 ...
#>  $ S        : num [1:100, 1:11] 0.0125 0.1691 0.028 0.1016 0.0537 ...
#>  $ Pg       : num [1:11, 1:11] 0.00 9.82e-09 0.00 0.00 0.00 ...
#>  $ Wg       : num [1:11, 1:11] 0.00 2.88e-01 1.13e-08 0.00 0.00 ...
#>  $ W        : num [1:11, 1:11] 0 0.445 0 0 0 ...
#>  $ Omega    : num [1:11, 1:11] 1.437 -0.339 -0.166 -0.289 -0.41 ...
#>  $ lowbound :'data.frame':	5 obs. of  5 variables:
#>   ..$ J        : num [1:5] -745 -743 -744 -744 -744
#>   ..$ T1       : num [1:5] -1028 -1068 -1076 -1077 -1078
#>   ..$ T2       : num [1:5] -8.03e-03 -9.08e-04 -1.36e-04 -2.67e-05 -6.34e-06
#>   ..$ T3       : num [1:5] 283 325 332 333 334
#>   ..$ parameter: chr [1:5] "complete" "complete" "complete" "complete" ...
#>  $ features :'data.frame':	4 obs. of  4 variables:
#>   ..$ diffPg   : num [1:4] 0.5 0.002725 0.000605 0.000112
#>   ..$ diffW    : num [1:4] 50413 707769 7316856 14584530
#>   ..$ diffOmega: num [1:4] 0.62287 0.05676 0.01324 0.00369
#>   ..$ diffWg   : num [1:4] 2.99e+10 4.90e+09 7.45e+10 7.79e+11
#>  $ finalIter: num 4
#>  $ time     : 'difftime' num 0.302571058273315
#>   ..- attr(*, "units")= chr "secs"
#>  $ max.prec : logi FALSE