Computes precision and recall statistics separating observed from hidden variables, and (FPR,FNR) for hidden variables .

Computes precision and recall statistics separating observed from hidden variables, and (FPR,FNR) for hidden variables .

ppvtpr(probs, G, r, thresh = 0.5)

Arguments

probs
matrix of estimated edges probabilities

G
original graph

r
number of missing actors

thresh
required threshold for criteria computations, default to 0.5

Value

PPV Precision of the whole data

PPVH Precision regarding observed data

PPVO Precision regarding hidden data

TPR Recall of the whole data

TPRH Recall regarding observed data

TPRO Recall regarding hidden data

FPRH False Positive Rate of hidden data

FNRH False Negative Rate of hidden data

Examples

data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- use true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(data$Y,cliqueList=initClique,sigma_O, MO,r=1 )
nestorFit=nestor(data$Y, MO,SO, initList=initList, maxIter=3,verbatim=1 )

#> 
#> nestor ran in 0.181secs and 3 iterations.

#-- obtain criteria
ppvtpr(nestorFit$Pg,r=1, data$G)

#> $PPV
#> [1] 1
#> 
#> $PPVH
#> [1] 1
#> 
#> $PPVO
#> [1] 1
#> 
#> $TPR
#> [1] 1
#> 
#> $TPRH
#> [1] 1
#> 
#> $TPRO
#> [1] 1
#> 
#> $FPRH
#> [1] 0
#> 
#> $FNRH
#> [1] 0
#>