Run function nestor on a list of initial cliques using parallel computation (mclapply)

Run function nestor on a list of initial cliques using parallel computation (mclapply)

List.nestor(cliqueList, Y, sigma_obs, MO, SO, r, alpha = 0.1,
  cores = 1, maxIter = 20, eps = 0.001, trackJ = FALSE)

Arguments

cliqueList
a list containing all initial cliques to be tested

Y
count dataset

sigma_obs
result of PLN estimation: variance covariance matrix of observed data

MO
result of PLN estimation: means matrix of observed data

SO
result of PLN estimation: marginal variances matrix of observed data

r
number of hidden variables

alpha
tempering parameter

cores
number of cores

maxIter
maximal number of iterations of the algorithm

eps
convergence precision parameter

trackJ
boolean for the estimation of the lower bound at each parameter update

Value

a list containing the fit of nestor for every cliques contained in cliqueList

Examples

 data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)

#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!

MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_obs=PLNfit$sigma_obs
#-- find a list of initial cliques
findcliqueList=boot_FitSparsePCA(data$Y, B=5, r=1)
cliqueList=findcliqueList$cliqueList
length(cliqueList)

#> [1] 1

#-- run List.nestor
List.nestor(cliqueList,data$Y, sigma_obs, MO,SO, r=1)

#> [[1]]
#> [[1]]$M
#>               [,1]        [,2]         [,3]         [,4]         [,5]
#>   [1,]  1.09935591  1.67535381 -1.482189164  1.123821947 -0.095125879
#>   [2,] -0.28264796 -0.04592392 -0.354144185  0.140249257  0.704654628
#>   [3,]  0.44231507  0.25082075  0.183879920  1.395826282 -0.220755775
#>   [4,] -0.24202104  0.29426118  1.268284087 -1.699428689 -1.461456362
#>   [5,] -1.10409625 -0.09902882 -0.321187431 -0.708059694 -0.580634140
#>   [6,] -0.70200932  0.38956924  0.364318459  0.197574075  0.019717967
#>   [7,]  1.31832381  1.32275156 -0.589047728  1.175025439  1.774074653
#>   [8,] -0.36276058  0.22736807  0.564019426 -0.361417234 -0.388661684
#>   [9,] -0.49513236 -0.50992076  0.357548219 -1.332095896 -0.768801186
#>  [10,]  0.18979393  2.08610439 -0.289194365 -0.018259143  0.923825260
#>  [11,] -2.11571579 -1.40001052  2.346261062 -1.522628444 -1.365275326
#>  [12,]  0.70766905  0.06966930 -1.047425671 -1.495175354 -0.452198524
#>  [13,]  1.67663577 -0.28021090 -0.562993855 -0.841911631  0.629759117
#>  [14,] -0.12390496 -0.75071288  0.097983736 -0.742576819 -0.570658409
#>  [15,] -0.53336482  0.68658483  0.151767103  0.588617485 -0.320330553
#>  [16,] -0.09156132 -0.13032498 -1.083273827  2.008664512  0.647259942
#>  [17,] -0.68946645 -0.56937170  0.285342456 -0.264154538 -0.421608713
#>  [18,]  1.93864137 -0.67132479 -1.473403844  0.088707773  1.279867256
#>  [19,]  0.96764539  0.46672317 -0.872399329 -0.965829550  0.667415604
#>  [20,]  0.76822032  0.06963943 -1.361091829  0.958580158  0.489261928
#>  [21,]  0.05859629 -0.72554830  0.325296143 -0.572862938 -0.922383035
#>  [22,]  0.80451839  0.22579281 -0.818139245  2.569703940  2.129436660
#>  [23,] -0.58005348  0.67284904  0.771316388  0.901926437  0.737926723
#>  [24,] -1.20162209 -1.47424413  2.392441106 -1.082435826 -2.085305803
#>  [25,] -1.84453814 -2.16537953  2.105591870 -0.671211395 -2.292201909
#>  [26,]  0.26380532  1.89491818 -1.262848338  1.627549610  0.440806258
#>  [27,]  0.21978060 -0.91032744 -0.069259603  0.534349597  0.165778030
#>  [28,] -1.15001923 -1.51619561  0.765579610 -0.668058659 -1.973541013
#>  [29,] -1.41288827 -0.35348223  2.481448290  0.060594836 -0.362050900
#>  [30,]  0.90023762  0.38610278 -0.904847960 -0.400299908  0.862430824
#>  [31,] -0.11445919 -1.01438811 -0.621741077 -0.624676659 -0.195682867
#>  [32,] -0.27123115 -0.20228767 -1.419250905  0.598590471  0.238587448
#>  [33,] -0.04156475  0.09605901  0.436871596 -0.014785294 -1.230931663
#>  [34,] -0.75451117  0.22354130  0.934884793  0.001479158  0.148587502
#>  [35,] -1.29773062 -1.39971794  0.326130700 -1.701960584 -0.104766001
#>  [36,]  1.50623304  0.63221008 -0.727985878  1.327220282  1.648204392
#>  [37,]  1.32168698 -0.17206598 -1.163310576  0.672509163  0.142569900
#>  [38,] -2.37114265 -0.83823603  3.653619827 -1.069879410 -1.456842706
#>  [39,]  0.49167329  0.70747287  0.025989890  0.054234187 -0.502565029
#>  [40,]  0.27764038 -0.94235964 -1.305380895 -1.369330782 -0.634002584
#>  [41,]  0.92417470 -0.27839257 -1.605571010  1.213948450  2.012469457
#>  [42,] -1.72130345 -0.94361118  2.417325537 -0.681742309  0.033635934
#>  [43,]  0.08473531  1.64896109 -0.472388701  0.080891917  0.466950658
#>  [44,] -1.79296821  0.40314866  3.301088490  0.434765116  0.161299933
#>  [45,]  1.18249320  0.72103983 -0.862111507  0.514296759  1.067919782
#>  [46,] -1.83033941 -0.41510097  3.100713254 -0.366795071 -0.661525885
#>  [47,]  0.77896561 -0.37530241 -0.059933874 -0.256210733  0.795938477
#>  [48,] -0.37432749 -0.54766350 -1.473004289 -0.019872664  0.156467741
#>  [49,] -0.67290122  0.53936303 -0.434458874 -0.513203761 -0.715218909
#>  [50,] -1.53462042 -1.28801617  3.586470078 -1.450349886 -0.627121055
#>  [51,] -0.33583453 -1.00926415  0.226638531  0.322075584 -0.005722303
#>  [52,]  1.07294051  1.24028484 -1.751340975 -0.851333354  1.885669498
#>  [53,]  1.78190040 -0.90670690 -1.384430723 -0.441166429 -0.006235490
#>  [54,]  0.19182316  0.32895983 -0.178118989 -0.220028927 -0.669761631
#>  [55,]  0.94114607 -0.48588880 -1.563532233  0.663163294  0.233967099
#>  [56,] -0.17031086 -1.60021996  0.229301685  0.738760988  0.104304771
#>  [57,]  1.16676202 -0.19202586  0.226998432  1.430993911 -0.521396006
#>  [58,] -0.29025402 -0.13916112  0.270667558  1.181326970 -1.414897223
#>  [59,] -0.48563881 -0.77501170 -1.217614708  0.422206676 -1.041173558
#>  [60,] -0.82688612  0.40893838  0.983855273 -1.244232544 -0.271748192
#>  [61,] -0.28480386 -0.37624549  0.963247153 -0.522250397 -0.506223889
#>  [62,] -0.77042745 -0.89998235  1.551677401 -1.444234172 -1.120814990
#>  [63,] -0.83107925  1.14275139  0.160299472  0.533926078  0.567870047
#>  [64,]  0.52452632  0.96313704 -1.014250837  1.379357910  1.396330368
#>  [65,] -0.56098857  0.37963328 -1.040268402 -0.173043962 -0.579474966
#>  [66,]  0.76800631  1.57471020 -1.922958978 -0.877633013 -0.955908134
#>  [67,]  0.10041480  0.60601856 -1.550141731 -1.490001946 -0.487846088
#>  [68,]  1.26356855 -1.37497756 -2.256589857 -0.722379139  1.697365576
#>  [69,]  0.70232041  0.67931139 -1.233117520  0.919121664  3.119786665
#>  [70,] -0.66462580 -0.09538827  1.029478634 -1.233574939 -0.263874097
#>  [71,] -1.19748153 -1.20963816  1.610689463  1.041601507 -1.460912552
#>  [72,]  0.56927480 -0.49059026 -0.120502052 -0.535552544  0.536159448
#>  [73,] -1.06968730 -0.31307908  1.217618493 -0.073423037 -1.028555073
#>  [74,]  1.35712794  0.26575388 -2.149787806 -0.497288518 -0.496588688
#>  [75,]  0.92828019  0.37576569 -1.555254696 -0.068612294  1.153339209
#>  [76,]  1.28167690  0.22534709 -1.204670087 -0.503076121 -0.971199794
#>  [77,]  1.29276758  1.17016248  0.651836108  1.094562593  1.377413098
#>  [78,] -1.78140994 -0.30434165  0.865039600 -1.918446410 -1.342520980
#>  [79,]  0.79427598 -0.08240347  0.636777194  1.859494732 -0.648167317
#>  [80,]  1.03793213  1.06205581 -1.241221304  0.447925536  2.157924113
#>  [81,]  1.36475288 -0.35133838 -1.197126681 -0.668253068 -0.636888074
#>  [82,]  0.02942679 -0.46830135 -0.371795989 -0.469232528 -0.904313394
#>  [83,]  1.04932596  2.02914958  1.309794115  2.024696288 -0.244873147
#>  [84,]  1.68565296 -0.10499232 -2.409774024  1.494830328  0.786375420
#>  [85,]  0.60077738  2.03741063 -0.790371123  2.237922742  1.090300330
#>  [86,]  0.48698019  2.13477766 -0.481998298 -0.124513001  0.295342381
#>  [87,]  0.65073943  0.22481569 -1.854490028  0.223359219  1.143520907
#>  [88,]  0.33986882  1.20803557 -0.786810849 -0.075077838  0.050105916
#>  [89,] -0.82868106 -0.92955805  0.042376255 -0.022438840  0.266780719
#>  [90,] -1.74798408 -0.50629614  1.904726186 -0.676028619 -0.743633505
#>  [91,] -0.13616576 -0.59078058 -0.374335239  0.298418251 -0.420592319
#>  [92,]  1.10100776 -0.13384436 -0.222973833  2.078091646  0.263509717
#>  [93,] -0.73162772  0.46576952  2.444216657 -0.231462442 -0.055895262
#>  [94,] -1.23299816 -1.13949526  1.012875003 -1.515603727 -0.928693054
#>  [95,]  0.09533106 -1.25632071 -0.403147300 -1.028889238 -0.905501900
#>  [96,]  0.31249979  0.30879611 -0.266321075 -0.515218519 -0.882882407
#>  [97,] -0.79091694 -0.31019057  1.058903364  0.455347520 -0.374320523
#>  [98,]  0.40375832  0.52309117  0.007470144  0.904488778  1.335545895
#>  [99,] -0.46463748  0.14057351  0.110274082 -0.686219208  0.509168337
#> [100,]  0.43420171 -0.11044923  0.090715021 -0.186629509  1.922208144
#>               [,6]        [,7]        [,8]        [,9]       [,10]       [,11]
#>   [1,] -1.08655743 -0.84797022 -0.03137225  0.78357100  1.84522559 -1.02911737
#>   [2,] -1.04250110  0.05449770  1.25061374  1.16437984  0.53961884 -0.22985927
#>   [3,]  0.03824479  0.58625162  0.61144761  0.34002463 -0.53024798  0.38572756
#>   [4,]  0.37457120 -0.16246129  1.10118782 -0.02399165 -1.48012970  0.68646452
#>   [5,] -0.86993703  2.27905381  2.28074730  1.26194594 -0.12024778  1.00789082
#>   [6,] -1.29036577  0.35070651  1.43433899 -1.49406630 -1.47309840  0.31995819
#>   [7,] -1.85987261  1.50503680  2.04080327  0.76997078  2.69423775  0.19239019
#>   [8,]  0.69201804  1.65724652 -1.50750882  3.13102577 -0.34231962  1.07599466
#>   [9,]  0.42087634 -0.06535630  0.95492340 -1.30301756  0.64647425  0.23545094
#>  [10,] -2.25908178 -0.73893967  2.00248655  0.97904052  2.86878145 -0.60363057
#>  [11,]  1.95687485  1.91909592  0.21625953 -1.41615038 -1.09157174  2.11551219
#>  [12,] -0.58635584 -0.36181019 -0.50244802 -1.29097526 -0.84152433 -0.56121525
#>  [13,] -1.01213136 -2.56008585  0.83324464 -1.72557368  0.92117771 -1.53190541
#>  [14,]  0.05684112 -0.29819721 -1.04766912  0.58941727 -0.85987853 -0.01492164
#>  [15,] -0.38911673  1.13936786  2.52601957 -0.25404317 -1.41829285  0.64441537
#>  [16,] -0.76008752  0.86863767  0.98429577 -0.72580905 -0.38775118 -0.15927664
#>  [17,]  1.52635826  1.24179540 -1.40655609 -0.53487762  1.64807387  0.76503381
#>  [18,]  0.04561580 -1.59908620 -0.92082033  0.70100326 -1.35835114 -1.57303191
#>  [19,] -0.49898348 -0.38952506 -0.97468651 -0.73917501  0.07417739 -0.65805769
#>  [20,]  0.44505324 -0.12173509 -0.70795264  2.18589547 -0.96743603 -0.72059844
#>  [21,]  0.04730421  0.36635709  1.81173055 -1.93638557 -1.66959631  0.44508182
#>  [22,] -0.09912636 -0.86374663 -0.16265885  2.95601849  0.96090647 -1.06521268
#>  [23,]  1.05927830  0.89899812 -0.53576466  0.00341930 -0.33866846  0.65088766
#>  [24,]  1.07317060 -0.96066521 -0.04839501 -2.32735660 -1.36670358  0.89313379
#>  [25,]  2.21027852  3.44725914 -2.41727428 -1.66075814 -2.68775459  2.85723074
#>  [26,] -1.94943765 -0.49166466  1.44339519  0.87716842 -0.09608819 -0.84366065
#>  [27,]  0.38190669  0.37053655 -0.92040725 -0.34438551  0.66349716  0.11919731
#>  [28,]  2.01466505  0.65855670 -2.40839750 -1.60899277 -0.35120659  0.92409992
#>  [29,] -0.36034211  0.28804437  0.72322268  0.44807941  0.25548423  1.26878911
#>  [30,] -0.65489439 -2.08042821  0.25619044  0.21484847  0.70280894 -1.49019798
#>  [31,] -0.23367345 -1.09717194  1.22848984 -1.45671586 -1.24881458 -0.75584637
#>  [32,]  0.58029694  0.76772437 -0.04805803 -0.31366958 -1.12892192 -0.29448989
#>  [33,]  0.66090723 -0.05048070 -0.32375768  0.16302637  0.74668919  0.34314108
#>  [34,] -1.40894397  1.14253322  2.11054383  0.11263465  0.04400620  0.92037198
#>  [35,]  0.69624358  1.61438177 -1.12632967 -0.96845414 -1.81966192  0.91157057
#>  [36,]  0.79392583 -1.06268037 -1.75950060  2.93545586  1.41889478 -1.04996742
#>  [37,] -1.42406664 -1.68305638  2.23787392  1.34039927 -0.03750777 -1.31423911
#>  [38,]  0.85924573  2.61668036 -1.08503958  0.17533764 -0.16532171  3.02482156
#>  [39,]  0.36414381 -0.35331173 -0.86947185  1.49187316  0.10746209 -0.08182239
#>  [40,]  0.63168190 -1.53075878 -1.53239070 -1.12445335 -0.66162382 -1.19390575
#>  [41,]  0.20170270 -0.86653851 -0.92867215 -1.40949156  0.28326526 -1.39324720
#>  [42,]  0.59775984  2.47889590 -0.83461098  0.96588262 -0.93378066  2.20786692
#>  [43,] -1.11151362 -1.40838860 -0.12737334  1.08527705  3.49262549 -0.93095169
#>  [44,] -1.10614149  2.68050199  1.65691146  0.41676698 -1.10564526  2.66928541
#>  [45,] -1.53661053 -0.53145927  1.45400914  0.70249949  1.98097755 -0.77720869
#>  [46,]  0.11908551  2.20746073  0.64623009  0.05779760 -1.61125447  2.47861621
#>  [47,] -1.00705063 -1.27918093 -1.47339846 -0.86126186 -1.06997293 -0.73775467
#>  [48,]  1.84364072 -0.75129099 -1.51341443  1.23966153 -0.22519237 -1.01594524
#>  [49,] -1.45836721  1.24158614  1.23625841  0.55222658 -1.44748061  0.49298664
#>  [50,]  0.76203843  0.35786452 -1.41385863 -1.85325050 -1.32995987  1.82134897
#>  [51,]  0.96640769  0.16704285 -2.09443712 -1.28670257 -0.45068160  0.17772746
#>  [52,] -0.71804723 -2.33082000 -0.72197721  1.30344460  3.31066766 -2.12287027
#>  [53,]  0.16156919 -2.10460481 -0.33897606 -1.95478034  1.06041766 -1.58636173
#>  [54,] -0.19424096  1.10882979  0.86738190  1.49493689  1.24607846  0.53626332
#>  [55,]  1.56447229 -1.61562818 -0.04711973 -1.37900159 -1.45862211 -1.47038669
#>  [56,]  1.63943665  0.83415032 -2.44974678 -0.67519306 -1.24279228  0.47484885
#>  [57,]  0.24799382 -1.17876266 -0.69924803  1.47768535  1.17659185 -0.37714213
#>  [58,]  1.82564107  1.85888670 -2.42283020 -1.25414078 -1.13796821  1.18912722
#>  [59,]  1.50116547  1.12338752 -0.86701454 -0.88655955  2.83426095  0.14280328
#>  [60,]  0.22672652  0.50934272  1.65389881  0.73729044 -1.45413992  0.70603131
#>  [61,]  0.88003875 -0.01310626  0.93562670  0.78188807 -1.49061360  0.48648523
#>  [62,]  1.43774126 -0.53038140  0.56091405 -1.43153610  0.21211541  0.58941485
#>  [63,] -1.56902913  0.69500166  1.82041673  0.87686364  0.09572938  0.31338864
#>  [64,] -0.80468566 -0.03494003  2.08035752  0.31913718  1.64381969 -0.65857290
#>  [65,]  0.41618437 -0.56914672 -1.09675458  0.19044087  0.61659596 -0.63675591
#>  [66,] -0.07941740 -1.40134719 -0.38235582 -1.15315998 -0.01846690 -1.35674905
#>  [67,] -0.94876377 -0.13446881 -0.82530977  1.19933241 -0.54087265 -0.66916785
#>  [68,]  2.48197203 -1.24561093 -2.42839520 -0.39138422  0.14009019 -1.80924016
#>  [69,] -1.73358948 -1.51376824  1.52049940  1.99810086 -1.55698397 -1.69168285
#>  [70,]  0.51882169 -0.12110733 -0.27826597  0.31610521 -0.09098158  0.43022643
#>  [71,]  0.76714320  1.13879934 -0.90692453 -1.89734152 -0.50043947  1.44375599
#>  [72,] -0.01937232  0.59376724  0.29707905 -0.28486475  1.23394907  0.14814220
#>  [73,]  0.32774701  0.15669933  1.29161827  0.80044930  1.41585109  0.75153664
#>  [74,] -0.17426884 -2.41835400  1.69658831 -1.34208414  0.86969237 -1.99662418
#>  [75,] -0.67944207 -1.14405258  1.70629885 -0.47092540  0.29141525 -1.37801104
#>  [76,] -0.05217205 -1.18605432  0.61937997 -1.93540654 -1.90526227 -0.94065401
#>  [77,] -1.34004104 -1.05349911  1.00053630  1.14277444  1.38260610 -0.40505713
#>  [78,] -0.12229369  1.65029516  1.35122922 -0.64900961 -2.27466573  1.34056361
#>  [79,]  0.52940434 -0.50018902 -1.60010105 -0.33282069  0.98653193  0.13680697
#>  [80,] -1.42285268 -0.71210862  1.32400876  0.32771474  1.05065663 -1.18298108
#>  [81,]  0.12365547 -1.45102155 -0.02961681  0.23426977 -0.57012181 -1.10901603
#>  [82,]  0.94284833  0.89773621 -1.61583164 -1.32154167 -0.65145161  0.38671827
#>  [83,] -1.15475271 -0.51220245 -0.01478841  1.19585457  0.75262563  0.36703836
#>  [84,] -0.41516618 -2.34473500 -1.07404690 -1.75007274 -1.33250047 -2.25924497
#>  [85,] -2.03300011 -1.44528246  1.19859169  0.90787307  2.72635920 -1.17610411
#>  [86,] -1.60640390  0.66095601  0.82482571  0.65206920  1.14778106  0.05632040
#>  [87,]  1.06044775  0.24066568 -1.73385464  0.75423462 -1.00594826 -0.86011841
#>  [88,]  1.00973099 -1.43190737 -2.06394088  0.09649125  0.32747578 -1.01943061
#>  [89,]  0.62157705  1.14373765 -0.16988491 -0.57451772 -0.93108449  0.51471546
#>  [90,]  0.14703920  2.89123039  0.41468724  1.04278033  1.47176335  2.28818132
#>  [91,]  0.33607667 -0.44036029 -0.15225239 -0.40144661 -0.41890870 -0.30884571
#>  [92,]  1.74782068 -0.56858968 -2.64016393 -0.80821004  0.30186808 -0.40064626
#>  [93,]  0.79326458  0.90997273 -0.47473147  0.38128152  0.06968431  1.49932917
#>  [94,]  0.03768417  0.40073950 -0.07378722  0.22485122 -2.24332039  0.76196542
#>  [95,]  0.48502469 -0.01178564 -1.90085463 -2.07061041 -0.97600472 -0.05176179
#>  [96,] -0.35111440 -1.44801983  1.19302849  0.16380064  1.45863813 -0.66639868
#>  [97,]  0.34959501  1.93322616 -1.81835567 -0.73306666  1.14923836  1.41875756
#>  [98,] -1.43599920 -0.37102577  1.18156364  0.51169990  0.31042040 -0.36124504
#>  [99,] -1.25798515 -0.40663692  1.06263681  1.64796292 -0.55011993 -0.21477241
#> [100,] -0.06522967 -1.19664138 -0.26713293  0.25326866  0.29163483 -0.79466636
#> 
#> [[1]]$S
#>              [,1]       [,2]        [,3]       [,4]        [,5]       [,6]
#>   [1,] 0.04408509 0.02636506 0.370021422 0.04636211 0.108055598 0.20307289
#>   [2,] 0.12400237 0.11168629 0.176096386 0.11021875 0.054834317 0.19818428
#>   [3,] 0.07512054 0.08964846 0.114495947 0.03598596 0.119291844 0.09602480
#>   [4,] 0.12092349 0.08670837 0.043517600 0.37867627 0.271279766 0.07343122
#>   [5,] 0.19075649 0.11596868 0.171761733 0.21049307 0.156065790 0.17937555
#>   [6,] 0.15750559 0.08049164 0.098325181 0.10506359 0.098527539 0.22603933
#>   [7,] 0.03644050 0.03639592 0.209588998 0.04422120 0.020273670 0.29031948
#>   [8,] 0.13020672 0.09127919 0.082585178 0.16382506 0.135601098 0.05618428
#>   [9,] 0.14068127 0.15216659 0.098822830 0.31183031 0.177951529 0.07066947
#>  [10,] 0.09050749 0.01791701 0.167571422 0.12539661 0.044998784 0.33230864
#>  [11,] 0.26481876 0.24171326 0.015479855 0.34609785 0.257408896 0.01762275
#>  [12,] 0.06103599 0.10271902 0.286067491 0.34110341 0.142161391 0.15011958
#>  [13,] 0.02642430 0.13127910 0.205658607 0.23049563 0.058609546 0.19478440
#>  [14,] 0.11213492 0.17549899 0.122977669 0.21555854 0.154905937 0.09463523
#>  [15,] 0.14375449 0.06326053 0.117594467 0.07515324 0.128784160 0.13146597
#>  [16,] 0.10977444 0.11852988 0.292711079 0.02005506 0.057710159 0.16777147
#>  [17,] 0.15647964 0.15782691 0.105108468 0.15222199 0.138977710 0.02647421
#>  [18,] 0.02075176 0.16768342 0.368290772 0.11501554 0.032390057 0.09549684
#>  [19,] 0.04928944 0.07570623 0.255206962 0.24992432 0.056685307 0.14165714
#>  [20,] 0.05811758 0.10272333 0.345753198 0.05395493 0.066303548 0.06925043
#>  [21,] 0.09923701 0.17300094 0.101671936 0.19139314 0.197045238 0.09535273
#>  [22,] 0.05643196 0.09139139 0.245949728 0.01160973 0.014389910 0.10658775
#>  [23,] 0.14753503 0.06398178 0.068662955 0.05680700 0.053223805 0.04064219
#>  [24,] 0.19875190 0.24920711 0.014802593 0.26896358 0.363586265 0.04013234
#>  [25,] 0.24735809 0.31364060 0.019570879 0.20517469 0.393458868 0.01381723
#>  [26,] 0.08579553 0.02147411 0.326597517 0.02889783 0.069156610 0.30017305
#>  [27,] 0.08857855 0.19154610 0.140932439 0.07882817 0.087369050 0.07297735
#>  [28,] 0.19456089 0.25343244 0.069014148 0.20473810 0.347129664 0.01667705
#>  [29,] 0.21562711 0.13777520 0.013567609 0.11762009 0.132921098 0.12878156
#>  [30,] 0.05215260 0.08071115 0.260771730 0.16863135 0.047584012 0.15701412
#>  [31,] 0.11144659 0.20216114 0.214561359 0.19857436 0.116987848 0.11771933
#>  [32,] 0.12313047 0.12456234 0.357377488 0.07449973 0.082196641 0.06182460
#>  [33,] 0.10620232 0.10073827 0.092261568 0.12503735 0.238450407 0.05770613
#>  [34,] 0.16185193 0.09154704 0.059208513 0.12340616 0.088633317 0.23955675
#>  [35,] 0.20650889 0.24168007 0.101579710 0.37913582 0.108903312 0.05597672
#>  [36,] 0.03083016 0.06617035 0.231254817 0.03837171 0.022869464 0.05142539
#>  [37,] 0.03634031 0.12200214 0.307522932 0.06978822 0.089065937 0.24128278
#>  [38,] 0.27926662 0.18427284 0.004262669 0.26679771 0.270630352 0.04854952
#>  [39,] 0.07234452 0.06217532 0.130458075 0.11822582 0.147510445 0.07405443
#>  [40,] 0.08495530 0.19480724 0.334794828 0.31844624 0.162095701 0.05916416
#>  [41,] 0.05112234 0.13111744 0.395231060 0.04265074 0.016114787 0.08447586
#>  [42,] 0.23875024 0.19494837 0.014445920 0.20664864 0.097417670 0.06090056
#>  [43,] 0.09744067 0.02701698 0.192456317 0.11573940 0.067610793 0.20585667
#>  [44,] 0.24380308 0.07962238 0.006051615 0.08593964 0.087684821 0.20517496
#>  [45,] 0.04102975 0.06145028 0.253374541 0.08021595 0.039431994 0.25411667
#>  [46,] 0.24638530 0.14337026 0.007373929 0.16439233 0.165261741 0.09015825
#>  [47,] 0.05762559 0.13975005 0.139879241 0.15129791 0.050529547 0.19427458
#>  [48,] 0.13110397 0.15575193 0.368213647 0.12562165 0.088051190 0.01963299
#>  [49,] 0.15511516 0.07138745 0.187121392 0.18336079 0.171538597 0.24519788
#>  [50,] 0.22499159 0.23029007 0.004557758 0.33296424 0.161312644 0.05286376
#>  [51,] 0.12811424 0.20161868 0.110459306 0.09464507 0.100577528 0.04416152
#>  [52,] 0.04509686 0.03919894 0.425786158 0.23194219 0.018212977 0.16343332
#>  [53,] 0.02398954 0.19119456 0.350442538 0.17386814 0.100592345 0.08721344
#>  [54,] 0.09037490 0.08440637 0.153669680 0.14706369 0.166210299 0.11438728
#>  [55,] 0.05040057 0.14990148 0.386646761 0.07036621 0.082513891 0.02554823
#>  [56,] 0.11558121 0.26177556 0.110177590 0.06580470 0.091935033 0.02381300
#>  [57,] 0.04160096 0.12368599 0.110433548 0.03480847 0.149546565 0.08136519
#>  [58,] 0.12465349 0.11925643 0.106481027 0.04396375 0.264605158 0.01997037
#>  [59,] 0.13991145 0.17792561 0.317892622 0.08686438 0.212549562 0.02710375
#>  [60,] 0.16782999 0.07927797 0.056616787 0.29644565 0.124085775 0.08280277
#>  [61,] 0.12419679 0.13983213 0.057701511 0.18457232 0.147903495 0.04767549
#>  [62,] 0.16316987 0.19049607 0.033327031 0.33186015 0.223280339 0.02875075
#>  [63,] 0.16820452 0.04275152 0.116771157 0.07884879 0.061894876 0.25780434
#>  [64,] 0.07053590 0.05003251 0.280056343 0.03654734 0.029036320 0.17244635
#>  [65,] 0.14598342 0.08112126 0.284841813 0.14174864 0.155939114 0.07094861
#>  [66,] 0.05813979 0.02893570 0.460690487 0.23596817 0.201381255 0.10505259
#>  [67,] 0.09642696 0.06762483 0.383955648 0.34015739 0.145931003 0.18790078
#>  [68,] 0.03823872 0.23917689 0.532322447 0.21257819 0.021823715 0.01061987
#>  [69,] 0.06131046 0.06365892 0.320783183 0.05592934 0.005445519 0.27639646
#>  [70,] 0.15444365 0.11564926 0.054299136 0.29455252 0.123338459 0.06511734
#>  [71,] 0.19842447 0.22221656 0.031494816 0.05001798 0.271276264 0.05264024
#>  [72,] 0.06811444 0.15036297 0.146817662 0.18636721 0.063653110 0.10038214
#>  [73,] 0.18796152 0.13417866 0.045608481 0.13111391 0.210905006 0.07628455
#>  [74,] 0.03522060 0.08863925 0.510091509 0.18125390 0.146870787 0.11273271
#>  [75,] 0.05094651 0.08138280 0.384937501 0.13062076 0.036442628 0.15952604
#>  [76,] 0.03763847 0.09141358 0.315377713 0.18184373 0.203355357 0.10289421
#>  [77,] 0.03727884 0.04172390 0.076409704 0.04763696 0.029556784 0.23160434
#>  [78,] 0.24299948 0.13340751 0.063103539 0.41954653 0.254256061 0.10842004
#>  [79,] 0.05690855 0.11460928 0.077435121 0.02315207 0.163718182 0.06453791
#>  [80,] 0.04645755 0.04590031 0.322136774 0.08495277 0.013997461 0.24114102
#>  [81,] 0.03497907 0.13759102 0.313915884 0.20455452 0.162424311 0.08986091
#>  [82,] 0.10124371 0.14828990 0.178491055 0.17754873 0.194762359 0.04509617
#>  [83,] 0.04599620 0.01891045 0.041865122 0.01974900 0.121543843 0.21070118
#>  [84,] 0.02620490 0.11644829 0.564397471 0.03277668 0.050964427 0.13384326
#>  [85,] 0.06644545 0.01876428 0.241424109 0.01605109 0.038628254 0.30910864
#>  [86,] 0.07259582 0.01710555 0.193820467 0.13653564 0.078342086 0.26204978
#>  [87,] 0.06387781 0.09144959 0.445701966 0.10286403 0.036778618 0.04059613
#>  [88,] 0.08112881 0.04034452 0.240840635 0.13129558 0.096114772 0.04248933
#>  [89,] 0.16800305 0.19350780 0.128727180 0.12581362 0.080270729 0.05968322
#>  [90,] 0.24064466 0.15183853 0.023763326 0.20586115 0.174927726 0.08818063
#>  [91,] 0.11303229 0.15987023 0.178831466 0.09653022 0.138865682 0.07576433
#>  [92,] 0.04401216 0.11881841 0.159178568 0.01875222 0.080477588 0.02149865
#>  [93,] 0.16005715 0.07576106 0.014071117 0.14844075 0.104727155 0.05144394
#>  [94,] 0.20130287 0.21499383 0.055133063 0.34482212 0.197868594 0.09605267
#>  [95,] 0.09676479 0.22704037 0.182764167 0.26013668 0.194908686 0.06699258
#>  [96,] 0.08278914 0.08574307 0.164617968 0.18341794 0.192053946 0.12800688
#>  [97,] 0.16485646 0.13392123 0.052848779 0.08440420 0.134150567 0.07495375
#>  [98,] 0.07736525 0.07234921 0.132454823 0.05668278 0.030746824 0.24254148
#>  [99,] 0.13823601 0.09746514 0.121716250 0.20729860 0.065161039 0.22236330
#> [100,] 0.07560814 0.11688960 0.123719364 0.14336166 0.017584516 0.10390681
#>               [,7]       [,8]        [,9]       [,10]     [,11]
#>   [1,] 0.195703644 0.11591112 0.078156498 0.021648539 0.2070077
#>   [2,] 0.096482616 0.03627475 0.055191512 0.074708006 0.2070077
#>   [3,] 0.060513613 0.06578780 0.115390862 0.187883995 0.2070077
#>   [4,] 0.115593899 0.04178612 0.156242613 0.373507504 0.2070077
#>   [5,] 0.012132128 0.01338954 0.050383723 0.133918590 0.2070077
#>   [6,] 0.074711808 0.03044873 0.429460689 0.371637890 0.2070077
#>   [7,] 0.025691561 0.01692707 0.079106724 0.009402347 0.2070077
#>   [8,] 0.022194795 0.34387664 0.008244433 0.161317092 0.2070077
#>   [9,] 0.106703517 0.04792301 0.385054717 0.067741012 0.2070077
#>  [10,] 0.180916346 0.01758000 0.065466250 0.007911333 0.2070077
#>  [11,] 0.017229161 0.09363485 0.411077535 0.287196244 0.2070077
#>  [12,] 0.135753018 0.17034432 0.382415800 0.239181127 0.2070077
#>  [13,] 0.494351221 0.05366726 0.485236171 0.052469528 0.2070077
#>  [14,] 0.129044991 0.25502802 0.092877450 0.242485605 0.2070077
#>  [15,] 0.036349928 0.01052107 0.187516382 0.358864935 0.2070077
#>  [16,] 0.046769571 0.04663166 0.265485348 0.167444560 0.2070077
#>  [17,] 0.033004930 0.32324426 0.231687117 0.026218064 0.2070077
#>  [18,] 0.316239498 0.23317785 0.084181067 0.344922043 0.2070077
#>  [19,] 0.138743965 0.24230114 0.268005828 0.113241167 0.2070077
#>  [20,] 0.111794907 0.19936755 0.020851939 0.262583330 0.2070077
#>  [21,] 0.073671145 0.02116824 0.537694564 0.420166718 0.2070077
#>  [22,] 0.197906657 0.12941723 0.009803857 0.050545040 0.2070077
#>  [23,] 0.045486482 0.17482782 0.152822263 0.160827259 0.2070077
#>  [24,] 0.211879027 0.11758007 0.635525391 0.346905725 0.2070077
#>  [25,] 0.003826595 0.55041035 0.469549962 0.706342050 0.2070077
#>  [26,] 0.150243663 0.03018440 0.071838669 0.131199747 0.2070077
#>  [27,] 0.073418506 0.23310101 0.200997585 0.066690553 0.2070077
#>  [28,] 0.056680769 0.54833644 0.456910716 0.162503324 0.2070077
#>  [29,] 0.078887983 0.05939477 0.105135904 0.096529570 0.2070077
#>  [30,] 0.405061873 0.09042383 0.128309962 0.064307626 0.2070077
#>  [31,] 0.232350589 0.03705041 0.420586095 0.320468647 0.2070077
#>  [32,] 0.051314968 0.11756751 0.196340156 0.294840084 0.2070077
#>  [33,] 0.105388080 0.14771342 0.134000678 0.061744962 0.2070077
#>  [34,] 0.036242617 0.01581718 0.139726786 0.116231288 0.2070077
#>  [35,] 0.023135355 0.26911226 0.312816512 0.459383545 0.2070077
#>  [36,] 0.227030186 0.39833449 0.010005079 0.032690958 0.2070077
#>  [37,] 0.331281129 0.01396484 0.046836357 0.124712164 0.2070077
#>  [38,] 0.008704847 0.26167261 0.132627010 0.139138764 0.2070077
#>  [39,] 0.134833660 0.22470531 0.040589802 0.110007985 0.2070077
#>  [40,] 0.304152531 0.34921306 0.345483322 0.208344444 0.2070077
#>  [41,] 0.198299225 0.23453207 0.409599971 0.094167311 0.2070077
#>  [42,] 0.009969344 0.21895181 0.066236244 0.256160815 0.2070077
#>  [43,] 0.283045364 0.12568432 0.059365263 0.004257052 0.2070077
#>  [44,] 0.008174369 0.02458405 0.107895591 0.290008195 0.2070077
#>  [45,] 0.154914750 0.02987807 0.084049658 0.018956431 0.2070077
#>  [46,] 0.013010793 0.06373407 0.146182779 0.405619209 0.2070077
#>  [47,] 0.261290407 0.33691538 0.291340152 0.282804529 0.2070077
#>  [48,] 0.182528786 0.34515705 0.051437048 0.146328251 0.2070077
#>  [49,] 0.033011394 0.03677651 0.095959075 0.365568628 0.2070077
#>  [50,] 0.074200982 0.32472124 0.516926598 0.338525905 0.2070077
#>  [51,] 0.087664020 0.47469072 0.381449774 0.176241954 0.2070077
#>  [52,] 0.452069789 0.20158128 0.048474111 0.005102234 0.2070077
#>  [53,] 0.409610589 0.14951837 0.542290621 0.046020360 0.2070077
#>  [54,] 0.037406768 0.05199298 0.040475429 0.038580441 0.2070077
#>  [55,] 0.319274130 0.11747837 0.402410231 0.368195881 0.2070077
#>  [56,] 0.048278214 0.55805157 0.256263252 0.319065645 0.2070077
#>  [57,] 0.245025942 0.19814688 0.041133726 0.041221147 0.2070077
#>  [58,] 0.018270524 0.55172575 0.374057933 0.296734396 0.2070077
#>  [59,] 0.036901239 0.22435860 0.296392351 0.008186565 0.2070077
#>  [60,] 0.064868743 0.02465747 0.081470587 0.367128237 0.2070077
#>  [61,] 0.102139606 0.04877925 0.078271477 0.376024911 0.2070077
#>  [62,] 0.154738384 0.06888098 0.414759539 0.100318441 0.2070077
#>  [63,] 0.054837548 0.02097528 0.071858199 0.111148148 0.2070077
#>  [64,] 0.104021667 0.01629731 0.117464896 0.026316763 0.2070077
#>  [65,] 0.159428219 0.26374621 0.130966831 0.069616487 0.2070077
#>  [66,] 0.281804705 0.15482368 0.351761798 0.122684873 0.2070077
#>  [67,] 0.112969891 0.21754077 0.053421629 0.189484054 0.2070077
#>  [68,] 0.255817311 0.55306388 0.208288426 0.106904173 0.2070077
#>  [69,] 0.301173295 0.02803465 0.025011024 0.392155833 0.2070077
#>  [70,] 0.111738512 0.14229325 0.117722451 0.130584687 0.2070077
#>  [71,] 0.036370362 0.23085472 0.527943860 0.183454275 0.2070077
#>  [72,] 0.060105260 0.08722376 0.192034537 0.039030173 0.2070077
#>  [73,] 0.088411878 0.03489102 0.076978226 0.032801544 0.2070077
#>  [74,] 0.468169218 0.02366547 0.393890043 0.055060393 0.2070077
#>  [75,] 0.239579356 0.02343978 0.221004430 0.093487361 0.2070077
#>  [76,] 0.246232840 0.06531166 0.537252950 0.482310058 0.2070077
#>  [77,] 0.225643674 0.04592584 0.056306463 0.033858442 0.2070077
#>  [78,] 0.022338655 0.03297299 0.251577025 0.585827825 0.2070077
#>  [79,] 0.151247257 0.36356845 0.199228600 0.049354570 0.2070077
#>  [80,] 0.177408868 0.03382897 0.116605896 0.046453707 0.2070077
#>  [81,] 0.290329564 0.11574516 0.126230522 0.193942756 0.2070077
#>  [82,] 0.045539859 0.36694482 0.389357363 0.206758037 0.2070077
#>  [83,] 0.152654960 0.11428949 0.053596031 0.061417929 0.2070077
#>  [84,] 0.454642522 0.25968497 0.491420773 0.338986720 0.2070077
#>  [85,] 0.289184873 0.03811233 0.069862380 0.009109337 0.2070077
#>  [86,] 0.056560590 0.05408509 0.087918241 0.042368907 0.2070077
#>  [87,] 0.082240623 0.39265832 0.080238599 0.270049341 0.2070077
#>  [88,] 0.287001015 0.46760703 0.141603561 0.090507045 0.2070077
#>  [89,] 0.036202196 0.13018126 0.238478911 0.255657690 0.2070077
#>  [90,] 0.006639119 0.07856853 0.061706129 0.031082943 0.2070077
#>  [91,] 0.144398289 0.12829906 0.209875849 0.171755114 0.2070077
#>  [92,] 0.159368875 0.60251140 0.280985254 0.092620864 0.2070077
#>  [93,] 0.045028227 0.16666813 0.111379844 0.113684726 0.2070077
#>  [94,] 0.071449698 0.12012286 0.127233871 0.576805662 0.2070077
#>  [95,] 0.102042345 0.43010216 0.571334308 0.264227899 0.2070077
#>  [96,] 0.289770538 0.03831502 0.133907614 0.031481067 0.2070077
#>  [97,] 0.016998937 0.41146191 0.266807550 0.042306998 0.2070077
#>  [98,] 0.136746018 0.03873140 0.099472987 0.091912146 0.2070077
#>  [99,] 0.140591968 0.04332767 0.034996241 0.190899918 0.2070077
#> [100,] 0.247926917 0.14097549 0.124222177 0.093468262 0.2070077
#> 
#> [[1]]$Pg
#>               [,1]         [,2]         [,3]         [,4]         [,5]
#>  [1,] 0.000000e+00 2.089353e-13 0.000000e+00 1.021665e-10 1.000000e+00
#>  [2,] 2.089353e-13 0.000000e+00 0.000000e+00 2.805846e-10 1.148416e-07
#>  [3,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 8.949519e-09
#>  [4,] 1.021665e-10 2.805846e-10 0.000000e+00 0.000000e+00 1.000000e+00
#>  [5,] 1.000000e+00 1.148416e-07 8.949519e-09 1.000000e+00 0.000000e+00
#>  [6,] 0.000000e+00 1.000000e+00 0.000000e+00 0.000000e+00 9.999999e-01
#>  [7,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [8,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [9,] 0.000000e+00 1.000000e+00 0.000000e+00 0.000000e+00 7.071080e-09
#> [10,] 0.000000e+00 1.000000e+00 0.000000e+00 2.112054e-16 1.791287e-16
#> [11,] 0.000000e+00 0.000000e+00 1.000000e+00 0.000000e+00 1.000000e+00
#>            [,6] [,7] [,8]        [,9]        [,10] [,11]
#>  [1,] 0.0000000    0    0 0.00000e+00 0.000000e+00     0
#>  [2,] 1.0000000    0    0 1.00000e+00 1.000000e+00     0
#>  [3,] 0.0000000    0    0 0.00000e+00 0.000000e+00     1
#>  [4,] 0.0000000    0    0 0.00000e+00 2.112054e-16     0
#>  [5,] 0.9999999    0    0 7.07108e-09 1.791287e-16     1
#>  [6,] 0.0000000    0    1 0.00000e+00 0.000000e+00     0
#>  [7,] 0.0000000    0    0 0.00000e+00 0.000000e+00     1
#>  [8,] 1.0000000    0    0 0.00000e+00 0.000000e+00     0
#>  [9,] 0.0000000    0    0 0.00000e+00 0.000000e+00     0
#> [10,] 0.0000000    0    0 0.00000e+00 0.000000e+00     0
#> [11,] 0.0000000    1    0 0.00000e+00 0.000000e+00     0
#> 
#> [[1]]$Wg
#>               [,1]         [,2]         [,3]         [,4]         [,5]
#>  [1,] 0.000000e+00 2.113070e-09 0.000000e+00 1.046666e-06 1.032503e+05
#>  [2,] 2.113070e-09 0.000000e+00 0.000000e+00 1.584163e-06 1.287572e-03
#>  [3,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 4.856398e-04
#>  [4,] 1.046666e-06 1.584163e-06 0.000000e+00 0.000000e+00 1.137319e+04
#>  [5,] 1.032503e+05 1.287572e-03 4.856398e-04 1.137319e+04 0.000000e+00
#>  [6,] 0.000000e+00 7.721774e+06 0.000000e+00 0.000000e+00 1.122802e+04
#>  [7,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [8,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [9,] 0.000000e+00 3.794882e+04 0.000000e+00 0.000000e+00 6.119832e-05
#> [10,] 3.232711e-13 3.244843e+04 0.000000e+00 1.015720e-12 1.492608e-12
#> [11,] 0.000000e+00 0.000000e+00 6.490176e+12 0.000000e+00 5.426434e+04
#>              [,6]         [,7]     [,8]         [,9]        [,10]        [,11]
#>  [1,]        0.00 0.000000e+00        0 0.000000e+00 3.232711e-13 0.000000e+00
#>  [2,]  7721774.09 0.000000e+00        0 3.794882e+04 3.244843e+04 0.000000e+00
#>  [3,]        0.00 0.000000e+00        0 0.000000e+00 0.000000e+00 6.490176e+12
#>  [4,]        0.00 0.000000e+00        0 0.000000e+00 1.015720e-12 0.000000e+00
#>  [5,]    11228.02 0.000000e+00        0 6.119832e-05 1.492608e-12 5.426434e+04
#>  [6,]        0.00 0.000000e+00 14250490 0.000000e+00 0.000000e+00 0.000000e+00
#>  [7,]        0.00 0.000000e+00        0 0.000000e+00 0.000000e+00 2.418686e+13
#>  [8,] 14250490.39 0.000000e+00        0 0.000000e+00 0.000000e+00 0.000000e+00
#>  [9,]        0.00 0.000000e+00        0 0.000000e+00 0.000000e+00 0.000000e+00
#> [10,]        0.00 0.000000e+00        0 0.000000e+00 0.000000e+00 0.000000e+00
#> [11,]        0.00 2.418686e+13        0 0.000000e+00 0.000000e+00 0.000000e+00
#> 
#> [[1]]$W
#>               [,1]         [,2]         [,3]         [,4]         [,5]     [,6]
#>  [1,] 0.000000e+00 6.926519e-09 0.000000e+00 4.329950e-06 9.932532e+04      0.0
#>  [2,] 6.926519e-09 0.000000e+00 0.000000e+00 8.344775e-06 5.714477e-03 189275.8
#>  [3,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 3.060566e-04      0.0
#>  [4,] 4.329950e-06 8.344775e-06 0.000000e+00 0.000000e+00 7.392416e+04      0.0
#>  [5,] 9.932532e+04 5.714477e-03 3.060566e-04 7.392416e+04 0.000000e+00  67506.8
#>  [6,] 0.000000e+00 1.892758e+05 0.000000e+00 0.000000e+00 6.750680e+04      0.0
#>  [7,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00      0.0
#>  [8,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 107740.8
#>  [9,] 0.000000e+00 1.160382e+05 0.000000e+00 0.000000e+00 2.462549e-04      0.0
#> [10,] 0.000000e+00 1.021333e+05 0.000000e+00 4.864789e-12 5.993383e-12      0.0
#> [11,] 0.000000e+00 0.000000e+00 2.793703e+05 0.000000e+00 3.896827e+04      0.0
#>           [,7]     [,8]         [,9]        [,10]     [,11]
#>  [1,]      0.0      0.0 0.000000e+00 0.000000e+00      0.00
#>  [2,]      0.0      0.0 1.160382e+05 1.021333e+05      0.00
#>  [3,]      0.0      0.0 0.000000e+00 0.000000e+00 279370.33
#>  [4,]      0.0      0.0 0.000000e+00 4.864789e-12      0.00
#>  [5,]      0.0      0.0 2.462549e-04 5.993383e-12  38968.27
#>  [6,]      0.0 107740.8 0.000000e+00 0.000000e+00      0.00
#>  [7,]      0.0      0.0 0.000000e+00 0.000000e+00 243983.06
#>  [8,]      0.0      0.0 0.000000e+00 0.000000e+00      0.00
#>  [9,]      0.0      0.0 0.000000e+00 0.000000e+00      0.00
#> [10,]      0.0      0.0 0.000000e+00 0.000000e+00      0.00
#> [11,] 243983.1      0.0 0.000000e+00 0.000000e+00      0.00
#> 
#> [[1]]$Omega
#>               [,1]       [,2]         [,3]        [,4]       [,5]       [,6]
#>  [1,]  1.332925138 -0.4022867  1.341211362 -0.40424564 -0.6661564  0.2848919
#>  [2,] -0.402286742  2.1036245  0.300477917 -0.38212512 -0.4482347  0.9871459
#>  [3,]  1.341211362  0.3004779  2.700622889  0.23046576  0.5305267 -0.2030890
#>  [4,] -0.404245638 -0.3821251  0.230465760  1.19604798 -0.4842342  0.2022434
#>  [5,] -0.666156356 -0.4482347  0.530526716 -0.48423423  2.0856186  0.4921098
#>  [6,]  0.284891912  0.9871459 -0.203089026  0.20224343  0.4921098  2.5049381
#>  [7,]  1.409918457  0.2609997 -0.857227702  0.11861755  0.4248074 -0.2246015
#>  [8,]  0.004541409 -0.4499917 -0.003941772 -0.02082548 -0.1592046  1.0873903
#>  [9,] -0.111372524 -0.5584397  0.065160843 -0.27511777 -0.4270913  0.3554013
#> [10,] -0.350509295 -0.5556623  0.223879716 -0.30872847 -0.3365012  0.3041388
#> [11,]  1.851083429  0.3426852 -2.143068101  0.22934153  0.6936777 -0.2749882
#>               [,7]         [,8]         [,9]      [,10]       [,11]
#>  [1,]  1.409918457  0.004541409 -0.111372524 -0.3505093  1.85108343
#>  [2,]  0.260999712 -0.449991740 -0.558439653 -0.5556623  0.34268521
#>  [3,] -0.857227702 -0.003941772  0.065160843  0.2238797 -2.14306810
#>  [4,]  0.118617546 -0.020825483 -0.275117766 -0.3087285  0.22934153
#>  [5,]  0.424807423 -0.159204590 -0.427091302 -0.3365012  0.69367771
#>  [6,] -0.224601467  1.087390295  0.355401277  0.3041388 -0.27498819
#>  [7,]  2.770559505  0.029199348 -0.003114342  0.2561980 -2.21482294
#>  [8,]  0.029199348  1.696836519 -0.193489022 -0.1108498  0.03234036
#>  [9,] -0.003114342 -0.193489022  1.249569775 -0.3250736  0.07679982
#> [10,]  0.256197966 -0.110849788 -0.325073567  1.2475029  0.28709766
#> [11,] -2.214822941  0.032340356  0.076799818  0.2870977  4.82627814
#> 
#> [[1]]$lowbound
#>            J        T1            T2       T3 parameter
#> 1  -621.5345 -823.4985 -1.074865e+02 309.4505  complete
#> 2  -612.5435 -938.3145 -4.512837e-01 326.2223  complete
#> 3  -607.4144 -954.3924 -1.823554e+00 348.8016  complete
#> 4  -609.6635 -963.7887 -1.535936e-01 354.2789  complete
#> 5  -610.0814 -966.6123 -8.001130e-02 356.6109  complete
#> 6  -610.0405 -967.3849 -5.267711e-02 357.3970  complete
#> 7  -609.9516 -967.4473 -6.507651e-02 357.5607  complete
#> 8  -609.9756 -967.4009 -4.756510e-02 357.4729  complete
#> 9  -609.9852 -967.2439 -3.251669e-02 357.2911  complete
#> 10 -609.8548 -966.8686 -2.843993e-02 357.0423  complete
#> 11 -609.6889 -966.4308 -9.590637e-03 356.7515  complete
#> 12 -609.6182 -966.1533 -4.708995e-03 356.5398  complete
#> 13 -609.8058 -966.3177 -1.620295e-03 356.5135  complete
#> 14 -610.1324 -966.8051 -1.146877e-04 356.6729  complete
#> 15 -610.4439 -967.3563 -1.015464e-04 356.9126  complete
#> 16 -610.6948 -967.8412  1.834709e-05 357.1464  complete
#> 17 -610.8867 -968.2310  2.572975e-05 357.3443  complete
#> 18 -611.0304 -968.5340 -8.591034e-06 357.5036  complete
#> 19 -611.1343 -968.7630 -1.000382e-06 357.6287  complete
#> 20 -611.2037 -968.9278 -1.473666e-06 357.7241  complete
#> 21 -611.2429 -969.0364 -1.976548e-08 357.7935  complete
#> 
#> [[1]]$features
#>          diffPg        diffW   diffOmega       diffWg
#> 1  1.771663e+00 3.134472e+01 6.537693437 4.914644e+32
#> 2  5.671131e-01 9.623481e+01 0.380313028 1.405909e+04
#> 3  1.837828e-01 6.613331e+02 0.407699324 7.979466e+15
#> 4  1.561325e-01 4.644288e+03 0.127905219 3.403847e+05
#> 5  1.001035e-01 3.492166e+04 0.044685426 1.033031e+16
#> 6  5.125613e-02 1.434813e+05 0.015378018 4.806362e+16
#> 7  5.714346e-02 1.139238e+06 0.008609276 1.757284e+17
#> 8  8.051521e-02 8.527083e+06 0.017706459 1.304007e+18
#> 9  1.074390e-01 6.628447e+07 0.024349063 8.390178e+18
#> 10 1.254638e-01 1.755688e+08 0.029193236 4.874097e+19
#> 11 1.127293e-01 1.000492e+09 0.044024033 5.863925e+19
#> 12 6.846952e-02 0.000000e+00 0.059686385 1.937258e+20
#> 13 2.702233e-02 5.254540e+07 0.067647400 0.000000e+00
#> 14 7.539884e-03 2.861056e+07 0.067934922 4.598933e+14
#> 15 1.694319e-03 0.000000e+00 0.064223937 5.541910e+14
#> 16 3.336639e-04 0.000000e+00 0.059659404 0.000000e+00
#> 17 6.234036e-05 0.000000e+00 0.055476026 0.000000e+00
#> 18 1.151210e-05 0.000000e+00 0.051757741 0.000000e+00
#> 19 1.842086e-06 0.000000e+00 0.048391682 7.602921e+12
#> 20 2.846755e-07 0.000000e+00 0.045276995 0.000000e+00
#> 
#> [[1]]$finalIter
#> [1] 20
#> 
#> [[1]]$time
#> Time difference of 1.25674 secs
#> 
#> [[1]]$max.prec
#> [1] FALSE
#> 
#> [[1]]$clique
#> [[1]]$clique[[1]]
#> [1] 3 7
#> 
#> 
#>