Run function nestorFit on a list of initial cliques using parallel computation (mclapply)

Run function nestorFit on a list of initial cliques using parallel computation (mclapply)

List.nestorFit(
  cliqueList,
  Y,
  sigma_O,
  MO,
  SO,
  r,
  alpha = 0.1,
  cores = 1,
  maxIter = 20,
  eps = 0.001,
  trackJ = FALSE
)

Arguments

cliqueList
a list containing all initial cliques to be tested

Y
count dataset

sigma_O
result of PLN estimation: variance covariance matrix of observed data

MO
result of PLN estimation: means matrix of observed data

SO
result of PLN estimation: marginal variances matrix of observed data

r
number of hidden variables

alpha
tempering parameter

cores
number of cores

maxIter
maximal number of iterations of the algorithm

eps
convergence precision parameter

trackJ
boolean for the estimation of the lower bound at each parameter update

Value

a list containing the fit of nestorFit for every cliques contained in cliqueList

Examples

 data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- find a list of initial cliques
findcliqueList=boot_FitSparsePCA(data$Y, B=5, r=1)
cliqueList=findcliqueList$cliqueList
length(cliqueList)

#> [1] 1

#-- run List.nestorFit
fitList=List.nestorFit(cliqueList,data$Y, sigma_O, MO,SO, r=1)
length(fitList)

#> [1] 1

str(fitList[[1]])

#> List of 12
#>  $ M        : num [1:100, 1:11] 1.101 -0.282 0.444 -0.24 -1.103 ...
#>  $ S        : num [1:100, 1:11] 0.0441 0.1241 0.0751 0.121 0.1909 ...
#>  $ Pg       : num [1:11, 1:11] 0.00 2.27e-10 1.00 3.20e-08 1.00 ...
#>  $ Wg       : num [1:11, 1:11] 0.00 1.11e-04 3.89e+09 1.55e-02 4.53e+06 ...
#>  $ W        : num [1:11, 1:11] 0.00 2.81e-04 1.18e+06 4.98e-02 3.48e+06 ...
#>  $ Omega    : num [1:11, 1:11] 3.26 -0.402 1.341 -0.404 -0.666 ...
#>  $ lowbound :'data.frame':	16 obs. of  5 variables:
#>   ..$ J        : num [1:16] -725 -679 -663 -562 -530 ...
#>   ..$ T1       : num [1:16] -1133 -1055 -1033 -924 -890 ...
#>   ..$ T2       : num [1:16] 97.938 -0.451 -0.345 -4.898 0.494 ...
#>   ..$ T3       : num [1:16] 309 377 370 367 359 ...
#>   ..$ parameter: chr [1:16] "complete" "complete" "complete" "complete" ...
#>  $ features :'data.frame':	15 obs. of  4 variables:
#>   ..$ diffPg   : num [1:15] 0.645 0.567 0.264 0.99 0.952 ...
#>   ..$ diffW    : num [1:15] 50.2 95.5 410.9 3184.4 37977.8 ...
#>   ..$ diffOmega: num [1:15] 6.539 1.708 0.301 1.01 0.867 ...
#>   ..$ diffWg   : num [1:15] 4.93e+32 4.93e+04 7.51e+15 4.33e+17 5.28e+19 ...
#>  $ finalIter: num 15
#>  $ time     : 'difftime' num 0.878084182739258
#>   ..- attr(*, "units")= chr "secs"
#>  $ max.prec : logi FALSE
#>  $ clique   :List of 1
#>   ..$ : int [1:2] 3 7